+++ srcs/ft_pitch_value.c
@@ -6,13 +6,13 @@
 /*   By: cledant <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/01/20 12:48:53 by cledant           #+#    #+#             */
-/*   Updated: 2016/01/20 22:47:19 by cledant          ###   ########.fr       */
+/*   Updated: 2016/07/28 21:30:57 by cledant          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <string.h>
 
-static double	ft_abs_db(double val)
+static float	ft_abs_db(float val)
 {
 	if (val < 0)
 		return (-val);
@@ -20,9 +20,9 @@ static double	ft_abs_db(double val)
 		return (val);
 }
 
-double			ft_pitch_value(double min, double max, size_t size)
+float			ft_pitch_value(float min, float max, size_t size)
 {
-	double		range;
+	float		range;
 
 	range = ft_abs_db(min - max);
 	return (range / size);
